+++ src/tests/loadtest.py
@@ -5,7 +5,7 @@ import time
 import pprint
 import hcpsdk
 
-T_NAMESPACE = 'n1.m.hcp2.snomis.local'
+T_NAMESPACE = 'n1.m.hcp73.archivas.com'
 T_DAAC      = 'n'
 T_DAAC_PWD  = 'n01'
 T_PORT      = 443